Move FleetsResultStorage tests to their own file
